SDK Updater: some refactoring and some new features.
The refactoring part involves moving as much as possible stuff
in SdkLib/internal/repository.

The UI has moved in SdkUiLib/internal/repository except a specific
public wrapper for calling the Sdk Updater window.

There are also a bunch of new classes to handle the internal
structures: Package is the base type and to match the XML element
names we have PlatformPackage, AddonPackage, DocPackage and
ToolPackage.

All headers have been fixed.