ADT: Tag classes with "@since GLE1"
This adds a @since GLE1 to all classes involved in the current GLE
and a @since GLE2 to the new classes. The idea is that both classes
are going to be mixed in the same packages and it will help later
to remove the old GLE by just searching for this specific tag.

The alternative would be to move the classes in a sub-package
layout.gle1 vs layout.gle2, but that seems unecessary.

Change-Id: I577cd81819db4c6e0209ebbd910a393fc2182619