ADT GLE: change error display to a separate sash.
This allows GEP to display both the latest error and the latest
successful rendering. The error still needs an icon and the
canvas will need to be grayed or something.

Also the LayoutCanvas will directly use the ILayoutResult and we
won't need to custom EditData in UiElementNode so the correspond
method is @deprecated (but obviously still supported for GLE1.)

Change-Id: Ia3afed836755dbd84a4511eaed0782c85a383ac9