Remove non-working high-res robot texture.
This was added with good intentions, as part of the general drive to
provide high-res textures.

Unfortunately, most OpenGL ES drivers require that a texture's dimensions
be a power of two on each side. The original texture was 128 x 128, and
someone helpfully created a 192 x 192 high res texture. But 192 is not
a power of two, so the hihg-res texture would not render.

We could create a 256 x 256 texture, which would work, but that's not
really needed for this particular texture, and we'd prefer if the
texture stayed small for rendering performance reasons.