Add a script to download and package the toolchain sources.
download-toolchain-sources.sh will download the current toolchain sources
from our git repositories then finely package them into a compressed
tarball that will later be used to rebuild the prebuilt binaries properly.

Due to the way our sources are stored in the git repository, the sources
in the tarball are laid out slightly differently than in the original toolchain
source archive used for NDK 1.6.

An upcoming patch will modify build-toolchain.sh to deal with the difference
and also allow the build of gcc-4.4.0 and gdb-6.8.