Update build-toolchain.sh to be able to build gcc 4.4.0
Update the script used to build the prebuilt toolchain binaries.
There are now two ways to use it:

1/ As usual, just invoke it and it will download the source tarball,
   uncompress it, configure the toolchain sources, build them, then
   package the result into a 'prebuilt' tarball.

2/ Generate a new source tarball with the 'download-toolchain-sources.sh'
   script, and invoke build-toolchain.sh with the --package=<tarball>
   option.

Additionally, you can now specify --gcc-version=4.4.0 to build GCC 4.4.0
(with ARMv7 / FPU support)

The option --gdb-version=6.8 has been added but the build doesn't seem
to work yet. This will probably be fixed in the toolchain git repository
though.

This will be used to refresh the NDK with gcc 4.4.0 prebuilts and support
for armeabi-v7a soon.