Backup/Restore sample application
A very simple application: one activity, with a few bits of persistent data that
are updated live by the activity.  To that are added a few alternative agent
implementations, to illustrate various approaches and tradeoffs when implementing
a backup/restore agent.  In particular, there are example agents that do everything
"by hand," either record-by-record or by directly backing up the app's entire
persistent data file; plus an example agent that illustrates how very easy it
is to implement the latter using the OS-provided helpers.

Part of bug #2545514

Change-Id: Iaca33a5113406360c23d6e2e59eb012f9f1e9400