Make runtest --path smarter when choosing path to build.
The command 'runtest --path <path to parent folder>'
is intended to find and build all tests contained within that parent folder.

However, the runtest logic was flawed in that it assumed building a parent
folder's Android.mk file would build all sub-directories.

This commit changes runtest to explictly look for the rule to build
sub-dir makefiles when creating the set of dirs to build.

Change-Id: Ia7f2172c1be79be8589be5f435fc4fb32569c994