Remove .classpath and .project files
During review of the patch set for native audio sample, someone
mentioned there should not be these files in git.  But we seem to
be inconsistent about this (for example these files).  Can you
please let me know whether they should be there or not, and then I
will update the native audio patch set. Thanks!

Change-Id: I6d27f4936f8c60efb05a89a4ba11e77e66aaf539