Add drag/drop demo to ApiDemos
Drag/drop among four big dots on screen.  Drags to a dot will have
the identity of the originating dot put into a text field below the dots.
In all cases, a text field to the right of the dots reports whether the
drag ended in a successful drop.

Attempting to start a drag from a dot reading "Drag ANR" should ANR and
be cleaned up properly.

Attempting to drop onto a dot reading "Drop ANR" should similarly ANR
and be cleaned up properly.

Drags from a dot labelled "Local" are restricted to targets within
the app's own window -- they are not draggable to the system "shirt
pocket" drop target, etc.

A drop onto the same dot that it originated from will append text to
that effect to the message that notes the dropped payload.  This
uses the "local state" convenience mechanism in startDrag() and
DragEvent.getLocalState().

Change-Id: Ic5cd6a29186a84c91d3dc4187e83e7bcf530ba2f