opengles emulator: support glEGLImageTargetRenderbufferStorageOES
That adds support to create a renderbuffer storage from an EGL
image (previously only texture was supported).
For that we have added Renderbuffer and Framebuffer attachment points
state tracking and when a renderbuffer which attached to an EGLImage
get attached to a framebuffer we attach the underlying EGLImage's
texture instead.

Also fixed some egl_image related bugs.
That makes the egl_image conformance test to pass on host.

Change-Id: I4e8c4935e45b27214038ecf1c1187ed96c5cea5b