opengles emulator: call XInitThreads in emulator_test_renderer
Now that we are not running the renderer as a seperate process
we need to have the XInitThreads call we put in the renderer process
inside emulator_test_renderer since the render server runs inside that
thread.
NOTE FOR INTEGRATION: this will need to be added to the emulator as
well.

Change-Id: I70592ad976d08c86e2c9e75b11b6ad57d6067606