Submit fixed script.
Not completely right - seems gubuntu/ubuntu both get unhappy after running
with the 'bridge' option and the PC's routes get hosed (right after the
script is run all is fine, but some agent then removes the routes on the
bridged iface and sets up routes on the native iface (eth0) and things on
the PC lose net access.  Note that you can manually reset the routes and make
it happy again).

The fix I applied to the script was related to turning on rndis on the device
and setting device side dns servers correctly so my 'fix' is not responsible
for the eventual unhappniess on the PC.  Also, the rev-tether continues
to work fine even when the PC is unhappy, so this may be useful.

Perhaps somebody can figure out who is changing the pc routes and why.

bug:4208971
Change-Id: I6ef6a88c71779b2083dad26a4be35ff74c26aae4