Update the api-versions.xml file for KitKat
This CL updates the api-versions.xml file, which is a file
used by Lint API check to check which API version each class,
field and method reference was introduced in.

It is generated by running sdk/apigenerator on an SDK install
which has android.jar from all 19 releases of Android installed.

This diff shows some methods getting removed from API 18 which
apparently were not in API 18; the last revision of the file
must have been run on a non-final version of the platform.

Change-Id: If08e4e9f08ae7fecbe938c1e928ad8698ed4025a