Update the api-version.xml file to reflect KitKat recent changes
This CL updates the api-versions.xml file, which is a file
used by Lint API check to check which API version each class,
field and method reference was introduced in.

It is generated by running sdk/apigenerator on an SDK install
which has android.jar from all 19 releases of Android installed.

This file was updated recently, but a few more APIs have changed
since then, so I regenerated it with the latest android.jar
snapshot.

Change-Id: I93b14fcd2e692aa607bbefade46f6f42ae144864