win_sdk: temporary disable ALL win_sdk build. DO NOT MERGE.
This deactivate building all the exe binaries
and prevents atree from running (both for dev.git
and sdk.git). This still results in a win_sdk folder
so that the sdk_repo rule can work, but the generated
sdk will only contain linux binaries and is not to be
used for production, obviously.

Change-Id: I99cd27a050278afae3b62d1ef34a0a1a542bfd1a