Fix SDK build.
development's sdk.atree depends on the generated file
/sdk/files/sdk_files_NOTICE.txt which was accumulating
all the tools notices.

However this file isn't generated anymore.
This replaces it by a static file.

This is only good as a temporary measure to fix the build.
The proper fix is to recreate the SDK rule to generate the
combined NOTICE.txt and move it in development.git

Change-Id: I27e4f38e756b81b002d16dfac5ccaa54a4463786