Add customtabs support lib to extras/android/support/
Currently the only way to get access to customtabs library through
download m2repository and using the aar file whereas all other
support libs are also present in extras/android/support/ with
the corresponding resources and jar file.

This adds the jar and manifest to the corresponding folder to avoid
the confusion to developers.

This is relanding
https://googleplex-android-review.git.corp.google.com/#/c/756492/
after it was reverted due to not having the corresponding Android.mk
change.

Change-Id: I21969994e1dd0bb4f9abbfa395cf284e3cfc396d