adb: fix device module for non-shell_v2 usage.
The device.py module was incorrectly querying the return code when the
shell_v2 return code isn't available.

This CL fixes it by specifying the return code probe as an argument
list rather than a single string argument. This means that before the
device was executing something like this:

  /system/bin/sh getprop '; echo $?'

which didn't do what we wanted. Now it does something like this:

  /system/bin/sh getprop ; echo $?

Bug: http://b/25470461
Change-Id: I5e20da31ec7ecc782c6146d8b38d752d52082860