Explicitly keep stubs as v51 / 1.7 class files
Without this change an associated change to the build
files will switch the Android stubs to v52 class files.
There are tools that are not quite ready yet. The target
will later be switched to "1.8" when the build is ready
and there are public classes using default / static
interface methods.

Bug: 26753820
Change-Id: I85e3f8ed0bb153af8ae0e83b0971b2ce408bf5b0