Fixed the video view crash in SupportLeanbackShowcase master build
The "Video consumption Examples" view in SupportLeanBackShowcase was failing on
the master build (fugu-userdebug N AAC30 2573265 devkeys),
since onCreate gets called in the reverse order of fragments' appearence in the
code when using a single transaction.
Fixed that by using separate transactions per fragment addition, and now the
example video view works on the master build as well as the LMP-MR1 build.

Bug: 26894030

Change-Id: Ic020cb87049af7e2095387ca1aab017f2b026630