SupportLeanbackShowcase: Added new credit card workflow and theme background for the wizard view
Created a payment workflow consisting of a set of fragments for
entering & adding new credit card information (checks in place
for validity of the card number as well as expiration date).
Also, changed the dropdown subaction's background colors to
be consistent with the rest of guided step fragment view.
Fixed the back key pressed to "not" exit the entire wizard on the last
fragment (it not goes back to the payment fragment section when back
button is pressed).

Change-Id: I39cf316f09b765cffbd0a7e012539c49dbe238ed