Showcase app: Fixed video fragment not showing video. Updated Settings and deprecated APIs for the app to be installable on L.
Video player was not being prepared for asynchronous playing after the
media source was set. Used the same preparation module for both the
Music and Video players.
Also, updated Settings to include a switch widget. Depredatecd some of
the APIs to level 1 to be able to install app on L.

Change-Id: I3bd8b266be028e80d87367a2c3c8a8174e924b92