KeyDerivationFunction: example about treating data encrypted via SHA1PRNG
The Crypto provider providing the SHA1PRNG algorithm for random number
generation was deprecated.

This algorithm was sometimes incorrectly used to derive keys.

This example provides a helper class and shows how to treat data that
was encrypted in the incorrect way and re-encrypt it in a proper way.

(cherry picked from commit 477c90ca1a08fa20448399c0606110e697d15667)

Bug: 27873296
Change-Id: I180d76c4aa7fb9aa8c332119bf094e050a63b58c