Backport machine/fenv.h to GB for ARM and mips.
Not needed for x86 because it has always had a working fenv
implementation.

We want to add the fixed implementations of fenv to libandroid_support
rather than dirtying up the bionic headers for these, but with the
inlines and the types in the same header we can't do this easily.
Backport machine/fenv.h for these architectures so we can simplify
this.

These are just copied from the android-21 directory. They aren't
moved because gen-platforms.sh actually removes all the currently
accumulated headers for android-21, so android-21 needs to have a
full set of headers...

Test: ndk/checkbuild.py
Bug: https://github.com/android-ndk/ndk/issues/120
Change-Id: Idc3bebc533230afca563c6911c65e90b51342a4b