vndk-def: Fix bad inward-customized vndk calculation.
This commit fixes bad inward customization vndk calculation.

Before this commit, when an vndk-indirect candidate is using
vndk-fwk-ext, its dependencies will be considered as a candidate of
vndk-indirect.

After this commit, the dependencies will only be pushed into the stack
if such dependencies are known to be intact or inward-customized.  If a
vndk-indirect candidate is extending APIs, then the dependencies of its
generic counterpart will be pushed into the stack.

This commit also fixes the ELF resolution so that vndk-core dependencies
will not be resolved to vndk-fwk-ext.

Test: ./tests/test_vndk.py
Change-Id: I0527ef62b70ca4442eb8f69272a8ba9a6388df55