vndk-def: Fix check-dep errors on vendor_libs
This commit fixes a check-dep bug introduced by 7e723126e4.  In
7e723126e4, graph.lib_pt[PT_VENDOR].values() was changed to return an
iterator instead of a set.  This means we can only traverse the iterator
once, thus vendor-to-vendor dependencies were treated as ineligible
dependencies.  This commit fixes the problem by creating a set() object
from the iterator, so that `vendor_libs` can be used several times.
This also makes it faster to lookup an object.

Test: Run `vndk_definition_tool.py check-dep` against sailfish images
and vendor-to-vendor dependencies should not show up.

Change-Id: I3b36efa635220941278c491b83f308a65e9f0ffe