vndk-def: normalize_partition_tags() should not rename files
This commit removes the rename_lib() from normalize_partition_tags().

Before this commit, a different heuristic was adopted.  If a system
shared library is not in the AOSP system image and it is depending on
some vendor libraries, then assume such library is a vendor shared
library that was incorrectly installed into system image.
normalize_partition_tags() will move such libraries into vendor
partition implicitly.  It turns out to be a confusing behavior to some
vendors.

This commit removes the aforementioned heuristic.  VNDK definition tool
will simply emit a warning and hide the violating dependencies in the
future.

Test: ./tests/run.py
Test: Run VNDK definition tool on sailfish images.
Test: Run VNDK definition tool on some internal images.
Change-Id: I760cfe22215829d757963f0532eb7f6427f72d84