Metalava: support names in APIs
This CL adds support for names in APIs: In Kotlin,
all parameter names are part of the API. In Java,
formal API parameter names can be specified with
@ParameterName("<name>"). Metalava now tracks this,
and in signature files, public parameter names
are included in signatures; in stubs, parameter
names (whether public or not) are always used but
the public name if specified will win.

This CL also adds enforcement of parameters in
signature file comparisons (you cannot remove or
change a parameter in an API), and when running
with --check-compatibility, all API change warnings
are turned into errors.

Some misc other bug fixes and cleanup.

Test: Unit tests included.
Change-Id: I0cfe0741f325812328cd643d51a2b1e9d644cc21