vndk-def: Fix is_zipfile() for Python 3.5
Python 3.5 adopts a fancy algorithm to detect the zip file.  It will
search for the magic words in the file, i.e. magic word does not have to
be at the beginning.  This commit adds an extra check to make sure "PK",
the magic word for zip file, is at the beginning of the file.

Test: ./vndk_definition_tool.py apk-deps --system ... --vendor ...
Change-Id: I3e633709bc39187ae807aef97c6e0396d02266ea