Prevent manifest file from being added to prebuilts
Change logic that handles the manifest file so that the manifest
is not added to the prebuilts/ directory but remains in a
temporary dir for GPL checking. At the end of the update.py script,
the manifest file is deleted when the temporary dir is deleted.

Also modify how check_gpl_license.py is run independently since
the manifest file has to be fetched first.

Test: python update.py -b oc-mr1-treble-dev --build 4728343 -vv 27
Bug: 77546642
Change-Id: I7749e2998cd12325b9c6275da5e697e328137b73
(cherry picked from commit a2504712decce9246c826923e0fcaa2b585ab349)