Add logic for replacing maps for zip files
For zip files, try to open the file in the current directory and
insert the files as mappings. This enables the script to pick up
embedded .so files.

Added a check to look in current directory before the symbol dir.

Added --app-symbols option to specify the directory containing the
app APK and so files.

Test: manual
Bug: 111268230
Change-Id: Id1b1f13bac78911daa6a72c35ebfadfdc9c1700f