Retry the command for querying parent artifacts
Currently mixed build targets are flaky, intermittently breaking
with the error message that either the parent artifact directory
does not exist (system_dir, device_dir) or a specific artifact
in the parent artifact directory does not exist.
To help debug this flakiness issue and avoid any race conditions
within the build-chaining infrastructure, allow the retrying of
the command for querying parent artifacts.

Test: build_mixed system_dir device_dir out_dir
Bug: 112250435
Change-Id: Ic2fd4056163c2a32d6057b6c182061c605cd158d