Dead code removal
Code to access the ActivityInstanceInfo api through reflection is no
longer needed, because we have the lifecycle tracking.

There were some resources for the snackbar that used to be shown during
the launching of the preset activities however that has been removed.

Finally there was a bunch of code in TestBase, that was trying to
predict the outcome of intent flags that never did anything.

Test: Build module
Change-Id: I73e0f172b605cff2433c8c386ef9b1672130f349