Winscope: Fix how layers with transforms and positions are rendered
Currently position is applied to the layer bounds before applying the transform. The bounds
are also not bounded by parent bounds.

Change the logic to use the calculated layer bounds and apply the transform with position when
rendering the layer on screen.
Keep existing logic to view traces before layer bounds were added.

Bug: 128925841
Test: open winscope with old (P) and new trace (Q ToT). Check if layers are rendered correctly when rotated.
Change-Id: Ia91262e4dfe3e8ad647acf8e0cf961e270441cde