Move lld (darwin, linux) and add dependent libraries
darwin: Move lld to lld-bin directory; add lib64/libc++.1.dylib; add trampoline

linux: Move lld to lld-bin directory; add lib64/libc++.so.1; add trampoline

Best practice is to ship necessary libraries of expected versions along
with build tools, rather than requiring them to be found elsewhere.

lld depends on libc++.1.dylib (darwin) or libc++.so.1 (linux) and looks for it
in ../lib64, so we need to move lld to an appropriate relative location.
In the old location of lld, we add a simple bash trampoline to invoke
lld in lld-bin; this means Gradle, which invokes lld in the old
location, does not have to change.

Bug: 148267171
Bug: 142590626
Bug: 144040336

Test: (gLinux) m TARGET_PRODUCT=sdk_phone_armv7 TARGET_BUILD_VARIANT=userdebug sdk dist sdk_repo
      (gLinux) m TARGET_PRODUCT=sdk_phone_armv7 TARGET_BUILD_VARIANT=userdebug win_sdk dist sdk_repo
Change-Id: I69b4ef79208bd7d06a377c0761671f1b572abced