Update the sample keyboard to showcase scroll, clip, toggle
The keyboard now showcases how one can implement a suggestion
strip with some pinned suggestions on the left and the right
of the strip and the strip suggestions are scrollable and well
as clipped to the scroll are's bounds. Also this showcases
how to toggle the layering of the suggestions to temporary
amimate on top of them. Fun, fun fun :)

bug:1234567

Change-Id: I7f2a6ea148de31be7cd0f13f56d72f6e5246a5f8