Store the sdump path for each type definition when linking ABI dumps
This commit adds sdump paths to ModuleIR::odr_list_map_. It maps type ID
to list of TypeDefinitions. A TypeDefinition consists of a TypeIR and
the sdump path where the type is defined. The path is used to
distinguish the definitions of the type in different source files.

When header-abi-linker merges modules, the compilation unit paths are
copied to the merged module. Thus, the paths are not lost when the
merged module is merged to another one.

Test: ./test.py
Bug: 147396457
Change-Id: I72a9502b4e81f8ead10c8439af76aea86e3bc3f3