Add functionality to search + navigate tags/errors.
Add a searchbar to bottom of Winscope that can be used to navigate through
all the tags/transitions/errors present in a trace (integrating existing timestamp search).
Choose the desired tab to set the type of search. Typing input into the relevant search will filter the elements.
Click on the start or end timestamp in the scrollable search table to navigate to that
timestamp in the timeline. Click on a transition in the search table to navigate to the start timestamp for that transition. Click on a transition in the timeline to toggle between start and end timestamps for that transition.

Bug: b/196197875

Test: upload the traces on the bug and test the above implementations on
the x20.

Change-Id: Ib3a0fff770cc65a38e50a8d99f130752b2f98eeb