Generate tags/errors for uploaded SF/ WM traces.
Tags can be generated after uploading SF/WM pair. If a Tag File is also
uploaded, clicking `Generate Tags` in the toolbar will regenerate the
tags, but not destroy the tag file, which can be redownloaded via
`Download All`. Otherwise, clicking `Generate Tags` will create tags
(which cannot be downloaded), and automatically display them where
relevant.

Bug: b/198590445
Test: upload a SF and WM trace without tags (example attached to bug),
click the button in the toolbar to Generate Tags and see that the tag
trace is generated and displays correctly. It should not download as a
trace if you click Download All.

Change-Id: I2c72de91691584add707065ac38ba83c49cfe705