Use TEST_MAPPING imports for third-party crate tests.
This script currently computes all transitive reverse dependencies for
a given crate and adds them to the generated TEST_MAPPING.  This
creates a very large file and makes it difficult to rename/remove
tests, as a given test can be in many TEST_MAPPING files in other
directories.  We fix this by instead importing the TEST_MAPPING files
of other third-party crates.

Note that we cannot do this for tests defined by this crate (as they
need to be in some TEST_MAPPING file) and for tests outside of
external.  For the latter, we do not auto-generate those TEST_MAPPING
files, so they might contain extra tests we do not want or not contain
the tests we do want.  We thus continue to list such tests
individually, which is suboptimal but at least is a strict improvement
over the current system.

Test: Run on a few crates.
Change-Id: Ibdf36e7844f5946c9e321ac60558792abd691b42