Try to locate symbols using build-id.
If the parsed backtrace has ELF build-ids, use them to
brute-force search for files in the symbols directory.

The search only considers files with the same basename,
so the number of candidates to check is very small.

This fixes issues when compile time directory does not
match the runtime directory on device (e.g. for apex).

Bug: 171980804
Test: ART crashes can be symbolized again
Change-Id: I0311a10c1e2eab92ee44f1b32c523d4763a54ee1