Winscope: Introduce ProxyClient class
Use new class to share device auth states, adb proxy
state and expose functions to talk to proxy.

This is the first step before adding new functionality to
winscope that will convert transaction traces to
layer traces. This functionality needs to be added in
DataInput.vue so this class will make it easier to read
the proxy state and use the proxy.

There should be no functional change in this cl.

Test: collect winscope traces via proxy
Bug: 200284593
Change-Id: Ic1f492ebec6c8dfd667d0828bb2d6779665b89c2