Remove e2fsdroid and sload_f2fs from platform sdk
fastboot used to reference these binaries as an option to populate newly
created filesystem. This option was not reachable from the command line
and was removed (commit 5d72d6c). Remove the unused binaries from the
platform sdk. These are still included in AOSP and can be built using
the regular platform build process [1].

[1] https://source.android.com/setup/build/building

Bug: 237960487
Test: m sdk
Change-Id: I1037c9e1ac7f76473a95676b93eb433fc37484e6