gdbclient.py: Fix gdb-remote command for non-local $ANDROID_SERIAL.
As it turns out, gdbclient.py is broken when connecting to a host:port
$ANDROID_SERIAL where the host is something other than localhost,
because it will try to connect to a TCP socket on the host named in
$ANDROID_SERIAL, while lldb-server is not even set up to listen on a TCP
socket on the device (it listens on a Unix socket which gets forwarded
by adb forward to localhost).

Fix it by entirely removing the code that tries to connect directly to
the host if the $ANDROID_SERIAL is of the form host:port.

Change-Id: I91a04aa811b246c9cac82ef3b0779dc284364edf