Replace logic for sdk source jar
Use the newly created srcjar for the bootclasspath as opposed to the
previous method of scanning the tree for *.java files.

The new srcjar har some known limitations:
- jarjar renaming / zapping is not taken into consideration
- optional classpath libraries (<library>) are not included in the
  sources at the moment.
- when prebuilts are used for bcp jars, their sources are not included

Overall it should still be a large improvement over the current jar
which contains a lot of irrelevant cruft, and is very incomplete.

Also delete the old script.

Bug: 151360309
Test: m sdk sdk_repo dist; inspect the sources.zip file
Change-Id: I1cf96966334e15c202a0ea38d6133581bb02b80c