Minor improvements to timeline input.
Allow any number of trailing zeros for nanosecond part of real time input of timeline. Arrow keys do not control timeline when timestamp inputs are focused.

Errors in timestamp input no longer propagate - the form control is used to indicate to the user if their input is valid or not (red line if invalid)

Bug: b/294349311 b/266923037
Test: npm run test:unit:ci
Change-Id: Idbcc69532ce91fed4b8eb8bcc1fe45a1b43df458