Update AndroidStyle.xml to match presubmit hooks.
This is based on me having imported the old style, then making changes to the style settings as I noticed discrepancies between the style and what the hooks expected. It's not perfect (there are some continuation indent differences in Kotlin in some edge cases, that IntelliJ doesn't have a separate setting for while ktfmt seems to treat it differently), but it's pretty close and definitely better than the current style.

Bug: 293451777

Change-Id: Iac895704c8fd4d7989795179b5756fdf3563a34e