Add test for MonkeySourceScript
This is the first test that will cover the behaviour of monkey command.
It's being added so that we can provably solve bugs in the monkey code
in the future.

It also highlights the incorrect behaviour of the current PinchZoom
command: it uses the wrong ACTION_POINTER_DOWN event, and it lacks an
ACTION_UP event.

Bug: 281806933
Test: atest monkey_test
Change-Id: Idad80fc0b2374978ba0a38a64c0de28259c389bc