build_vendor: Add argument for release config
To let this script work with trunk stable branches, we need to
add an argument for the release configuration.

Temporarily only, we make this default to trunk_staging if the
argument is not provided.  However (b/310040467), we will soon
make this a required argument and break if it is not provided.

Bug: 310040467, 307946156
Test: TreeHugger
Change-Id: Ic7996e9436b5fc4456ca7f8b30098c5821790faa