Winscope: Use factory method pattern to create new UiData objects
The existing pattern in Winscope viewers is to create a static UiData
object called EMPTY, and use a reference to that directly when a new
object is needed. This re-uses the static object rather than creating a
new one. Refactor the existing pattern to use a factory method instead,
which can help prevent bugs resulting from the reuse/misuse of the EMPTY
object.

Bug: 332714237
Bug: 210460522
Test: Presubmit
Flag: NONE dev tool only
Change-Id: I154109c2a89eab4dd11f02e6353a6b3585a9778e