Winscope: Update TransformMatrix labels to to match platform code
The labels used for different coefficients within the TransformMatrix
class used in Winscope do not match the labels for those values from the
platform. In this CL, we update Winscope's use of these labels to match
the labels in the platform. Functionally, there should be no behavior
change to the Winscope UI/tool.

This is needed because the transform coefficients were written incorrectly from
the platform initially. Since then, a new method to write the transform values
correctly has been introduced, but it can be only used for newly added
transforms to ensure backwards compatibility.

To account for this, we adjust the transform node directly to correct
labels in the cases where we know the transform was written
incorrectly.

Bug: 332714237
Bug: 210460522
Test: Presubmit
Flag: EXEMPT dev tool only
Change-Id: I82dd59184ae35db7357f4b6e47bad36531a8c6f7