Add Android.bps for google_metadata and rooted_path.
So we can run tests at presubmit, with the Android build system.

Requires adding documentation for public methods to satisfy the build
system, so do that as well.

Also requires some cleverness to make protos work with both Soong and
Cargo.

Also requires using UTC rather than the local timezone because the
chrono crate in external/rust/crates doesn't enable the "clock" feature.
Should be good enough, since we are just using it to get the date.

Change-Id: Ie9a1c2101e3e9a30f791a141bda8cd965ef35ccb
Bug: 339424309
Test: treehugger