cargo2rulesmk.py: Pick rust version from envsetup.sh
The current strategy of picking the latest Rust version under prebuilts
can fail if it is newer than what is supported or tested by Trusty. If
we pick a Rust prebuilt version from envsetup.sh in
trusty/vendor/google/aosp/scripts, running cargo to build the crate
is more likely to succeed.

Test: manually, from path/to/zerocopy, run cargo2rulesmk.py --run
      --config cargo2rulesmk.json
Bug: None
Change-Id: I47c7bf37192cb135b8092d3e71e482a1689f7597