Make fetchartifact more progress bar friendly.
We need a way to report the content length of the download for progress
bars. This could have alternately been done with passing callbacks or by
returning the content length and a future that the caller separately
awaits, but this seemed the simplest, and preserves behavior of
fetch_artifact_chunked for existing callers.

Bug: None
Test: pytest
Change-Id: Ib898aa48555d858f0dff812559690c18fc5ca3f4