Improve handling of size during overlay interruptions
This CL prevents the size of the demo overlays from instantly jumping to
their target value when interrupting the overlay transitions, for
instance when interrupting a QSShade => NotificationShade transition by
a NotificationShade => Lockscreen transition.

The trick used in this CL and the reason why it is necessary is
explained in details in b/290930950#comment22.

Bug: 308961608
Test: Manual, triggered the above interruption in the demo
Flag: EXEMPT demo app
Change-Id: Ifa76b1519d77d434edc65e083027a6b26c259b67