Support deleting directories from crates.
The reason is that some crates, particularly *-sys crates, often include
the source for the libraries they are providing a wrapper for. These
libraries may have different, potentially unacceptable license terms,
and are not needed for Android, because they are imported and maintained
separately in the source repo.

It would be nice if we could do this with patch files, but that doesn't
work well because there is no way to say "delete this whole directory,
regardless of contents". You have to specify the full contents of every
file.

For now, the only crate we will delete from is libbpf-sys, so we can
migrate it to the monorepo.

This should probably be made configurable in the future, but isn't for
now.

Bug: http://b/339424309
Test: treehugger
Change-Id: Iaf140791239e537de4a7839c4b8bf4957a6a1913