Make the STL demo shades be completely separate elements
As we are still discussing with UX about what a shared model on handheld
could look like, there was a consensus that always going for a strictly
separate shade model is the best way to move forward in the short term.
This is applied to the STL demo to help us make it look more like the
production dual shade.

Bug: 308961608
Test: Manual, used the STL demo app
Flag: EXEMPT demo app
Change-Id: I86cc047886f87518ee28780f7c2c663bffc38063