Introduce SharedElement transitions with NestedSTLs
This is a first iteration to support shared elements across nestedSTLs.
This version has several limitations that may be addressed in future
CLs. Known limitations are:
- Interruptions are not supported (will jump cut)
- Side-by-side NestedSTLs are not supported. That means if we have a
ParentSTL that contains NestedSTL1 and NestedSTL2 then a shared element
can be transitioned between the ParentSTL and the NestedSTLs but not
between NestedSTL1 and NestedSTL2. Multiple nesting layers are supported
though, so ParentSTL > NestedSTL1 > NestedNestedSTL1 would support
sharing between any of them.
- If the sharedElement appears more than two times on the screen
(due to multiple transitions of multiple STLs transitioning at the same
time) the behavior is not supported.
- A sharedElement can't be present twice in the same hierachy. Meaning
that ParentSTL sceneA contains an element. There can't be a NestedSTL
on the same scene that would contain the same element in any of its
scenes. This would result in a conflict of the idle state.
- Currently the sharedElement is always rendered in the ParentSTL during
transitions. We don't support more control for this behavior as we do
for normal sharedElements yet.
- If you disable a sharedElement transition for a specific transition
you can only reference the element in the parentSTL to define a
custom transition. This is a general limitation where you can't
reference elements within NestedSTLs in transition DSL.
- NestedSTls are not allowed to share the same contentkeys

Test: Manual test app added + Units tests added
Bug: b/376659778
Flag: NONE not in production yet
Change-Id: Iffe44b2b77155ac94bd265872ffa78d9f5ba8cdb