Do not exempt deleted functions and global variables
header-abi-diff reports ABI difference for the functions and global
variables whose declarations are deleted, but the symbols are unchanged.
This case happens when a developer modifies the availability attributes
annotated on the declarations.

"-allow-adding-removing-referenced-apis" disables the check. The flag is
needed for C runtime libraries where not all functions are implemented
in C.

Test: ANDROID_BUILD_TOP=`realpath .` \
      PATH=$PATH:`realpath out/soong/dist/bin` \
      development/vndk/tools/header-checker/tests/test.py
Bug: 342516369
Change-Id: I2aafe37450994f3e8527f7980fb6753d295e9a80