Elimate "staging" as a thing separate from regenerating.
Previously, because of the needs of the crate migration code, we had a
notion of "staging" a crate, i.e. copying the crate code and android
customizations to a temporary directory in preparation for running
cargo_embargo. The details of what we did were slight different between
regenerating a crate in the monorepo vs. checking the migratability of a
legacy crate in external/rust/crates.

Now that the migration code is gone, we can simplify the logic and get
rid of the notion of "staging" a crate.

Test: ./crate_tool regenerate --all
Change-Id: I566ac467d313897e17dd30b6d3aea3752be8fbb4