Refactor handling of symlinks in checksum files.
We have a single case of a symlink to a file in a different crate:
crates/log/src/android_logger.rs

This presents a problem if we want to do all the work of
regenerating a crate in a temporary directory, and never
in place (which we do), because the symlink will dangle
while we are running regenerate, until we replace the old
crate contents.

To handle this:

1) Omit symlinks from the checksum file that point outside the crate directory.
Otherwise, resolve them as before and compute the checksum of the file they point to.

2) When verifying, always treat dangling symlinks as an error.

Unfortunately, we have to skip symlinks entirely in the AyeAye analyzer, so our
only chance is to validate locally.

Test: ./crate_tool regenerate uguid log
Change-Id: Ic79712ac720d4591b40908693cd12c28fe96e001