Be more thorough and correct about updating METADATA when regenerating.
Original impetus was a bug where we weren't updating the license terms
in METADATA when updating a crate, and there was at least one case where
the license terms changed.

In addition to fixing that, simplify METADATA update logic to just
expose a single update() method, and have init() just call update().

Also refactor ManagedCrate to have another state, indicating that we
have copied and patched the vendored code, and it is therefore safe to
find the license files, since we now need access to them not only to set
up symlinks and MODULE_LICENSE_* files, but also for updating METADATA.

Test: ./crate_tool regenerate --all
Change-Id: I96396e257147722198ffbc7a7192632130de77c4