repo-pull: allow passing arguments to merge or pick
To have more extensive usage of pulling changes, allow passing arguments
to the git command specified by the `--merge` or `--pick` options.

For example, if there is a change which is already existing in the
project, the script should fail currently. However, if the user pass
`--pick-args="--empty=drop"`, the script will simply ignore the change.

Along with this change, update README.md file to indiciate how to use
newly added `--merge-args` and `--pick-args`.

Bug: 417933507
Test: Verified in one project that trying pulling a change which already
  exists in the project with `--pick-args="--empty=drop"` works
Change-Id: I970268272a9edc657e9746dcee48a70923303dbb
Signed-off-by: Junghoon Jang <junghoonjang@google.com>