Properly update transform on change in rect.
We cannot guarantee that the matrix of a rect mesh has not been changed
under the hood from when applyMatrix4 was called, so simply applying the
inverse of the last applied transform does not guarantee that the rect
will be brought to its untransformed state in layer space.

When transforms change we should completely redraw the rect.

Fixes: 436228275
Test: npm run test:unit:ci
Change-Id: I597fd3f2fdfb9f6f2c19e2be3a24f5f72fdf8f77